% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastpolicytree-package.R
\docType{package}
\name{fastpolicytree-package}
\alias{fastpolicytree-package}
\title{fastpolicytree: Constructs policy trees from covariate and reward data}
\description{
A package for constructing optimal policy trees from
    covariate and reward data. It provides a subset of the
    functionality of the policytree package but constructs the
    trees more quickly (particularly for discrete data) and thus
    allows e.g. deeper trees to be constructed.
}
\keyword{internal}
